refactor: style header columns
