+++ src/utils/data.ts
@@ -105,3 +105,22 @@ export const userAppliedInfo = {
   buttonText: "View your submitted application",
   buttonLink: "forms.google.com",
 };
+
+export const customizeDataGridStyles = {
+  border: 'none',
+  '& .MuiDataGrid-row': {
+    borderBottom: '1px solid #000000',
+  },
+  '& .MuiDataGrid-columnHeader': {
+    backgroundColor: '#CCE4F0',
+    fontWeight: 'bold',
+    fontSize: '18px',
+    border: 'none',
+  },
+  '& .MuiDataGrid-columnSeparator': {
+    display: 'none',
+  },
+  '& .MuiDataGrid-cell:focus': {
+    outline: 'none',
+  },
+}
